now printing all important values into serial + fixes + value tweaking

+++ include/sensors/mpu.h
@@ -8,9 +8,9 @@ void recordAccelRegisters();
 
 void recordGyroRegisters();
 
-void getAccForce(float& aX, long& aY, long& aZ);
+void getAccForce(float& aX, float& aY, float& aZ);
 
-void getGyroRotation(float& gX, long& gY, long& gZ);
+void getGyroRotation(float& gX, float& gY, float& gZ);
 
 void getAccRaw(long& aX, long& aY, long& aZ);
